+++ src/IdeatecAPI.Domain/Entities/Proveedor/CompraProveedor.cs
@@ -6,7 +6,8 @@ public class CompraProveedor
     public int SucursalId { get; set; }
     public int ProductoId { get; set; }
     public decimal? PrecioCompra { get; set; }
-    public int? Cantidad { get; set; }
+    // Decimal: se pueden comprar 30.5 kg, 12.75 L, etc.
+    public decimal? Cantidad { get; set; }
     public string? UnidadMedida { get; set; }
     public DateTime? FechaCreacion { get; set; }
     public string? DocReferencia { get; set; }
